+++ src/main.js
@@ -1,6 +1,7 @@
 import { createApp } from 'vue'
 import App from './App.vue'
+import routers from './routers'  // router import
 
-
-
-createApp(App).mount('#app')
+const app = createApp(App)
+app.use(routers)  // router 추가
+app.mount('#app')
