+++ src/views/TestComponent.vue
@@ -1,0 +1,16 @@
+<template>
+  <div>
+    <h2>Test Component</h2>
+    Test Page
+  </div>
+</template>
+
+<script>
+export default {
+  name: "TestComponent"
+}
+</script>
+
+<style scoped>
+
+</style>
